Moving code to broader headers to allow for modularity
