+++ bin/echo.c
@@ -42,9 +42,9 @@ void print_s2n_error(const char *app_error)
 }
 
 /* Poll the given file descriptor for an event determined by the blocked status */
-static int wait_for_event(int fd, s2n_blocked_status blocked)
+int wait_for_event(int fd, s2n_blocked_status blocked)
 {
-    struct pollfd reader = { .fd = fd, .events = 0 };
+    struct pollfd reader = {fd, 0};
 
     switch (blocked) {
     case S2N_NOT_BLOCKED:
@@ -81,8 +81,8 @@ int early_data_recv(struct s2n_connection *conn)
     ssize_t total_data_recv = 0;
     ssize_t data_recv = 0;
     bool server_success = 0;
-    s2n_blocked_status blocked = 0;
-    uint8_t *early_data_received = malloc(max_early_data_size);
+    s2n_blocked_status blocked = (s2n_blocked_status)0;
+    uint8_t *early_data_received = (uint8_t*)malloc(max_early_data_size);
     GUARD_EXIT_NULL(early_data_received);
 
     do {
@@ -93,7 +93,7 @@ int early_data_recv(struct s2n_connection *conn)
 
     if (total_data_recv > 0) {
         fprintf(stdout, "Early Data received: ");
-        for (size_t i = 0; i < total_data_recv; i++) {
+        for (size_t i = 0; i < (size_t)total_data_recv; i++) {
             fprintf(stdout, "%c", early_data_received[i]);
         }
         fprintf(stdout, "\n");
@@ -106,7 +106,7 @@ int early_data_recv(struct s2n_connection *conn)
 
 int early_data_send(struct s2n_connection *conn, uint8_t *data, uint32_t len)
 {
-    s2n_blocked_status blocked = 0;
+    s2n_blocked_status blocked = (s2n_blocked_status)0;
     ssize_t total_data_sent = 0;
     ssize_t data_sent = 0;
     bool client_success = 0;
@@ -119,25 +119,8 @@ int early_data_send(struct s2n_connection *conn, uint8_t *data, uint32_t len)
     return S2N_SUCCESS;
 }
 
-int negotiate(struct s2n_connection *conn, int fd)
+void print_connection_data(struct s2n_connection *conn,  bool session_resumed)
 {
-    s2n_blocked_status blocked;
-    while (s2n_negotiate(conn, &blocked) != S2N_SUCCESS) {
-        if (s2n_error_get_type(s2n_errno) != S2N_ERR_T_BLOCKED) {
-            fprintf(stderr, "Failed to negotiate: '%s'. %s\n",
-                    s2n_strerror(s2n_errno, "EN"),
-                    s2n_strerror_debug(s2n_errno, "EN"));
-            fprintf(stderr, "Alert: %d\n",
-                    s2n_connection_get_alert(conn));
-            S2N_ERROR_PRESERVE_ERRNO();
-        }
-
-        if (wait_for_event(fd, blocked) != S2N_SUCCESS) {
-            S2N_ERROR_PRESERVE_ERRNO();
-        }
-    }
-
-    /* Now that we've negotiated, print some parameters */
     int client_hello_version;
     int client_protocol_version;
     int server_protocol_version;
@@ -177,31 +160,32 @@ int negotiate(struct s2n_connection *conn, int fd)
     printf("Curve: %s\n", s2n_connection_get_curve(conn));
     printf("KEM: %s\n", s2n_connection_get_kem_name(conn));
     printf("KEM Group: %s\n", s2n_connection_get_kem_group_name(conn));
+    printf("Cipher negotiated: %s\n", s2n_connection_get_cipher(conn));
+
+    if (session_resumed) {
+        printf("Resumed session\n");
+    }
 
     uint32_t length;
     const uint8_t *status = s2n_connection_get_ocsp_response(conn, &length);
     if (status && length > 0) {
         fprintf(stderr, "OCSP response received, length %u\n", length);
     }
+}
 
-    printf("Cipher negotiated: %s\n", s2n_connection_get_cipher(conn));
-
-    bool session_resumed = s2n_connection_is_session_resumed(conn);
-    if (session_resumed) {
-        printf("Resumed session\n");
-    }
-
+void psk_early_data(struct s2n_connection *conn,  bool session_resumed)
+{
     uint16_t identity_length = 0;
     GUARD_EXIT(s2n_connection_get_negotiated_psk_identity_length(conn, &identity_length), "Error getting negotiated psk identity length from the connection\n");
     if (identity_length != 0 && !session_resumed) {
-        uint8_t *identity = malloc(identity_length);
+        uint8_t *identity = (uint8_t*)malloc(identity_length);
         GUARD_EXIT_NULL(identity);
         GUARD_EXIT(s2n_connection_get_negotiated_psk_identity(conn, identity, identity_length), "Error getting negotiated psk identity from the connection\n");
         printf("Negotiated PSK identity: %s\n", identity);
         free(identity);
     }
 
-    s2n_early_data_status_t early_data_status = 0;
+    s2n_early_data_status_t early_data_status = (s2n_early_data_status_t)0;
     GUARD_EXIT(s2n_connection_get_early_data_status(conn, &early_data_status), "Error getting early data status");
     const char *status_str = NULL;
     switch(early_data_status) {
@@ -212,6 +196,31 @@ int negotiate(struct s2n_connection *conn, int fd)
     }
     GUARD_EXIT_NULL(status_str);
     printf("Early Data status: %s\n", status_str);
+}
+
+int negotiate(struct s2n_connection *conn, int fd)
+{
+    s2n_blocked_status blocked;
+    while (s2n_negotiate(conn, &blocked) != S2N_SUCCESS) {
+        if (s2n_error_get_type(s2n_errno) != S2N_ERR_T_BLOCKED) {
+            fprintf(stderr, "Failed to negotiate: '%s'. %s\n",
+                    s2n_strerror(s2n_errno, "EN"),
+                    s2n_strerror_debug(s2n_errno, "EN"));
+            fprintf(stderr, "Alert: %d\n",
+                    s2n_connection_get_alert(conn));
+            S2N_ERROR_PRESERVE_ERRNO();
+        }
+
+        if (wait_for_event(fd, blocked) != S2N_SUCCESS) {
+            S2N_ERROR_PRESERVE_ERRNO();
+        }
+    }
+
+    bool session_resumed = s2n_connection_is_session_resumed(conn);
+
+    print_connection_data(conn, session_resumed);
+
+    psk_early_data(conn, session_resumed);
 
     printf("s2n is ready\n");
     return 0;
